refactor(docs): chip-dx
