+++ apps/docs/content/components/chip/avatar.raw.jsx
@@ -1,0 +1,20 @@
+import {Chip, Avatar} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex gap-4">
+      <Chip
+        avatar={<Avatar name="JW" src="https://i.pravatar.cc/300?u=a042581f4e29026709d" />}
+        variant="flat"
+      >
+        Avatar
+      </Chip>
+      <Chip
+        avatar={<Avatar getInitials={(name) => name.charAt(0)} name="JW" size="sm" />}
+        variant="flat"
+      >
+        Avatar
+      </Chip>
+    </div>
+  );
+}
