+++ apps/docs/content/components/chip/close-button.ts
@@ -1,15 +1,4 @@
-const App = `import {Chip} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-4">
-      <Chip onClose={() => console.log("close")}>Chip</Chip>
-      <Chip onClose={() => console.log("close")} variant="bordered">
-        Chip
-      </Chip>
-    </div>
-  );
-}`;
+import App from "./close-button.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
