+++ apps/docs/content/components/chip/colors.ts
@@ -1,17 +1,4 @@
-const App = `import {Chip} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-4">
-      <Chip color="default">Default</Chip>
-      <Chip color="primary">Primary</Chip>
-      <Chip color="secondary">Secondary</Chip>
-      <Chip color="success">Success</Chip>
-      <Chip color="warning">Warning</Chip>
-      <Chip color="danger">Danger</Chip>
-    </div> 
-  );
-}`;
+import App from "./colors.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
