+++ apps/docs/content/components/chip/custom-styles.raw.jsx
@@ -1,0 +1,15 @@
+import {Chip} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Chip
+      classNames={{
+        base: "bg-gradient-to-br from-indigo-500 to-pink-500 border-small border-white/50 shadow-pink-500/30",
+        content: "drop-shadow shadow-black text-white",
+      }}
+      variant="shadow"
+    >
+      New
+    </Chip>
+  );
+}
