+++ apps/docs/content/components/chip/disabled.raw.jsx
@@ -1,0 +1,9 @@
+import {Chip} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Chip isDisabled color="primary">
+      Chip
+    </Chip>
+  );
+}
