+++ apps/docs/content/components/chip/radius.ts
@@ -1,15 +1,4 @@
-const App = `import {Chip} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-4">
-      <Chip radius="full">Full</Chip>
-      <Chip radius="lg">Large</Chip>
-      <Chip radius="md">Medium</Chip>
-      <Chip radius="sm">Small</Chip>
-    </div>  
-  );
-}`;
+import App from "./radius.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
