+++ apps/docs/content/components/chip/sizes.raw.jsx
@@ -1,0 +1,11 @@
+import {Chip} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex gap-4">
+      <Chip size="sm">Small</Chip>
+      <Chip size="md">Medium</Chip>
+      <Chip size="lg">Large</Chip>
+    </div>
+  );
+}
