+++ apps/docs/content/components/chip/usage.raw.jsx
@@ -1,0 +1,5 @@
+import {Chip} from "@nextui-org/react";
+
+export default function App() {
+  return <Chip>Chip</Chip>;
+}
