+++ apps/docs/content/components/chip/variants.ts
@@ -1,18 +1,4 @@
-const App = `import {Chip} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-4">
-      <Chip color="warning" variant="solid">Solid</Chip>
-      <Chip color="warning" variant="bordered">Bordered</Chip>
-      <Chip color="warning" variant="light">Light</Chip>
-      <Chip color="warning" variant="flat">Flat</Chip>
-      <Chip color="warning" variant="faded">Faded</Chip>
-      <Chip color="warning" variant="shadow">Shadow</Chip>
-      <Chip color="warning" variant="dot">Dot</Chip>
-    </div> 
-  );
-}`;
+import App from "./variants.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
